refactor: 폴더 이름을 더 직관적인 명칭으로 변경
