+++ src/features/friend-search/components/card/index.ts
