+++ src/features/friend-search/ui/section/index.ts
